Move eventmesh to dal
