+++ api-controllers/ServiceBrokerApiController.js
@@ -7,7 +7,7 @@ const errors = require('../common/errors');
 const utils = require('../common/utils');
 const catalog = require('../common/models/catalog');
 const FabrikBaseController = require('./FabrikBaseController');
-const lockManager = require('../eventmesh').lockManager;
+const lockManager = require('../data-access-layer/eventmesh').lockManager;
 const AssertionError = assert.AssertionError;
 const BadRequest = errors.BadRequest;
 const PreconditionFailed = errors.PreconditionFailed;
