+++ api-controllers/ServiceFabrikApiController.js
@@ -7,7 +7,7 @@ const jwt = require('../broker/lib/jwt');
 const logger = require('../common/logger');
 const backupStore = require('../data-access-layer/iaas').backupStore;
 const filename = backupStore.filename;
-const eventmesh = require('../eventmesh');
+const eventmesh = require('../data-access-layer/eventmesh');
 const lockManager = eventmesh.lockManager;
 const errors = require('../common/errors');
 const FabrikBaseController = require('./FabrikBaseController');
