+++ broker/lib/fabrik/UnlockResourcePoller.js
@@ -1,10 +1,10 @@
 'use strict';
 
 const pubsub = require('pubsub-js');
-const eventmesh = require('../../../eventmesh');
+const eventmesh = require('../../../data-access-layer/eventmesh');
 const CONST = require('../../../common/constants');
 const logger = require('../../../common/logger');
-const lockManager = require('./../../../eventmesh').lockManager;
+const lockManager = eventmesh.lockManager;
 const errors = require('../../../common/errors');
 const NotFound = errors.NotFound;
