+++ broker/lib/index.js
@@ -5,10 +5,10 @@ exports.store = require('./store');
 exports.middleware = require('./middleware');
 exports.fabrik = require('./fabrik');
 exports.bootstrap = bootstrap;
-// const config = require('../../common/config');
-// if (config.enable_swarm_manager) {
-//   exports.docker = require('./docker');
-// }
+const config = require('../../common/config');
+if (config.enable_swarm_manager) {
+  exports.docker = require('../../data-access-layer/docker');
+}
 const logger = require('../../common/logger');
 const docker = exports.docker;
 
@@ -19,5 +19,4 @@ function bootstrap() {
     .tap(() => logger.debug('Successfully fetched docker images:'))
     .spread((images) => images.forEach(image => logger.debug(image.status)))
     .catch((err) => logger.error('Failed to bootstrap docker client', err));
-
 }
