+++ broker/lib/middleware.js
@@ -9,7 +9,7 @@ const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
 const utils = require('../../common/utils');
 const catalog = require('../../common/models/catalog');
-const lockManager = require('../../eventmesh').lockManager;
+const lockManager = require('../../data-access-layer/eventmesh').lockManager;
 const CONST = require('./../../common/constants');
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
 const UnprocessableEntity = errors.UnprocessableEntity;
