+++ data-access-layer/eventmesh/ApiServerClient.js
@@ -3,12 +3,12 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const assert = require('assert');
-const config = require('../common/config');
-const logger = require('../common/logger');
-const CONST = require('../common/constants');
+const config = require('../../common/config');
+const logger = require('../../common/logger');
+const CONST = require('../../common/constants');
 const kc = require('kubernetes-client');
 const JSONStream = require('json-stream');
-const errors = require('../common/errors');
+const errors = require('../../common/errors');
 const BadRequest = errors.BadRequest;
 const NotFound = errors.NotFound;
 const Conflict = errors.Conflict;
