+++ data-access-layer/eventmesh/LockManager.js
@@ -3,10 +3,10 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const eventmesh = require('./');
-const errors = require('../common/errors');
+const errors = require('../../common/errors');
 const DeploymentAlreadyLocked = errors.DeploymentAlreadyLocked;
-const logger = require('../common/logger');
-const CONST = require('../common/constants');
+const logger = require('../../common/logger');
+const CONST = require('../../common/constants');
 const NotFound = errors.NotFound;
 const Conflict = errors.Conflict;
 const assert = require('assert');
