+++ managers/BaseManager.js
@@ -5,7 +5,7 @@ const Promise = require('bluebird');
 const errors = require('../common/errors');
 const logger = require('../common/logger');
 const config = require('../common/config');
-const eventmesh = require('../eventmesh');
+const eventmesh = require('../data-access-layer/eventmesh');
 const CONST = require('../common/constants');
 const NotImplementedBySubclass = errors.NotImplementedBySubclass;
 const Conflict = errors.Conflict;
