+++ managers/backup-manager/BackupService.js
@@ -8,7 +8,7 @@ const errors = require('../../common/errors');
 const bosh = require('../../data-access-layer/bosh');
 const backupStore = require('../../data-access-layer/iaas').backupStore;
 const utils = require('../../common/utils');
-const eventmesh = require('../../eventmesh');
+const eventmesh = require('../../data-access-layer/eventmesh');
 const Agent = require('../../broker/lib/fabrik/Agent');
 const ScheduleManager = require('../../jobs');
 const CONST = require('../../common/constants');
