+++ managers/backup-manager/DefaultBackupManager.js
@@ -2,7 +2,7 @@
 
 const Promise = require('bluebird');
 const catalog = require('../../common/models/catalog');
-const eventmesh = require('../../eventmesh');
+const eventmesh = require('../../data-access-layer/eventmesh');
 const logger = require('../../common/logger');
 const CONST = require('../../common/constants');
 const bm = require('./');
