+++ test/test_broker/support/index.js
@@ -42,4 +42,4 @@ global.chai.use(require('chai-http'));
  * Loading it from the first time before the test starts and mocking it here so that the tests need not do it.
  */
 mocks.apiServerEventMesh.nockLoadSpec();
-require('../../../eventmesh').apiServerClient.init();
+require('../../../data-access-layer/eventmesh').apiServerClient.init();
